feat:add filter by team size
