+++ src/modules/student/features/team-offers/useAddTeamOffer.js
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import toast from "react-hot-toast";
-import addTeamOffer from "../api/apiAddTeamOffer";
+import addTeamOffer from "../../api/apiAddTeamOffer";
 import { useNavigate } from "react-router";
 
 export function useAddTeamOffer() {
