+++ src/modules/student/features/team-offers/useDeleteCurrentLeaderTeamOffer.js
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import toast from "react-hot-toast";
 import { useNavigate } from "react-router";
-import { deleteCurrentLeaderTeamOffer } from "../api/apiStudent";
+import { deleteCurrentLeaderTeamOffer } from "../../api/apiStudent";
 
 export function useDeleteCurrentLeaderTeamOffer() {
   const navigate = useNavigate();
